Improves Investors page responsiveness
Enhances the responsiveness of the Investors page for better mobile viewing.

Adapts font sizes, paddings, margins, and layout of components such as the calculator, FAQ section, and header to ensure a consistent and user-friendly experience across different screen sizes.

Adds logout functionality to the NavBar.

Updates the API endpoint for profile-related requests.