Enables fetching lead data for portal
Adds endpoints to retrieve opportunity documents, legal contracts, and SIFCO numbers for a lead based on email or DPI.
These endpoints are protected by the portal token.

Also modifies the public lead creation endpoint to return the existing lead instead of an error if a lead with the same email, phone, or DPI already exists.